+++ src/plugins/targetinfo/MQ2TargetInfo.cpp
@@ -85,7 +85,7 @@ bool bDisablePluginDueToBadUI = false;
 bool gBUsePerCharSettings = false;
 bool gBShowDistance = true;
 bool gbShowPlaceholder = true;
-bool gbShowRoleplaying = true;
+bool gbShowAnon = true;
 bool gbShowSight = true;
 bool gbShowTargetInfo = true;
 
@@ -398,7 +398,7 @@ void HandleINI(eINIOptions Operation)
 		gBShowDistance = GetPrivateProfileBool(szSettingINISection, "ShowDistance", gBShowDistance, INIFileName);
 		gbShowTargetInfo = GetPrivateProfileBool(szSettingINISection, "ShowTargetInfo", gbShowTargetInfo, INIFileName);
 		gbShowPlaceholder = GetPrivateProfileBool(szSettingINISection, "ShowPlaceholder", gbShowPlaceholder, INIFileName);
-		gbShowRoleplaying = GetPrivateProfileBool(szSettingINISection, "ShowRoleplaying", gbShowRoleplaying, INIFileName);
+		gbShowAnon = GetPrivateProfileBool(szSettingINISection, "ShowAnon", gbShowAnon, INIFileName);
 		gbShowSight = GetPrivateProfileBool(szSettingINISection, "ShowSight", gbShowSight, INIFileName);
 
 		DistanceLabelToolTip = GetPrivateProfileString(szSettingINISection, "DistanceLabelToolTip", DistanceLabelToolTip, INIFileName);
@@ -426,7 +426,7 @@ void HandleINI(eINIOptions Operation)
 		WritePrivateProfileBool(szSettingINISection, "ShowDistance", gBShowDistance, INIFileName);
 		WritePrivateProfileBool(szSettingINISection, "ShowTargetInfo", gbShowTargetInfo, INIFileName);
 		WritePrivateProfileBool(szSettingINISection, "ShowPlaceholder", gbShowPlaceholder, INIFileName);
-		WritePrivateProfileBool(szSettingINISection, "ShowRoleplaying", gbShowRoleplaying, INIFileName);
+		WritePrivateProfileBool(szSettingINISection, "ShowAnon", gbShowAnon, INIFileName);
 		WritePrivateProfileBool(szSettingINISection, "ShowSight", gbShowSight, INIFileName);
 
 		WritePrivateProfileString(szSettingINISection, "DistanceLabelToolTip", DistanceLabelToolTip, INIFileName);
@@ -652,7 +652,7 @@ void ShowHelp()
 	WriteChatf("     \ay/targetinfo distance [%sOn\ay|%sOff\ay]\aw will toggle showing distance to target.", gBShowDistance ? "\ag" : "", gBShowDistance ? "" : "\ag");
 	WriteChatf("     \ay/targetinfo info [%sOn\ay|%sOff\ay]\aw will toggle showing detailed target info.", gbShowTargetInfo ? "\ag" : "", gbShowTargetInfo ? "" : "\ag");
 	WriteChatf("     \ay/targetinfo placeholder [%sOn\ay|%sOff\ay]\aw will toggle showing placeholder/named info.", gbShowPlaceholder ? "\ag" : "", gbShowPlaceholder ? "" : "\ag");
-	WriteChatf("     \ay/targetinfo roleplaying [%sOn\ay|%sOff\ay]\aw will toggle showing roleplaying/anon in the target display.", gbShowRoleplaying ? "\ag" : "", gbShowRoleplaying ? "" : "\ag");
+	WriteChatf("     \ay/targetinfo anon [%sOn\ay|%sOff\ay]\aw will toggle showing anon/roleplaying in the target display.", gbShowAnon ? "\ag" : "", gbShowAnon ? "" : "\ag");
 	WriteChatf("     \ay/targetinfo sight [%sOn\ay|%sOff\ay]\aw will toggle showing O/X based on your line of sight to target.", gbShowSight ? "\ag" : "", gbShowSight ? "" : "\ag");
 	WriteChatf("     \ay/targetinfo reset\aw will reset all settings to default.");
 	WriteChatf("     \ay/targetinfo reload\aw will reload all settings.");
@@ -689,10 +689,10 @@ void CMD_TargetInfo(SPAWNINFO* pPlayer, char* szLine)
 		gbShowPlaceholder = GetBoolFromString(szArg1, !gbShowPlaceholder);
 		WriteIni = true;
 	}
-	else if (ci_equals(szArg1, "roleplaying"))
+	else if (ci_equals(szArg1, "anon"))
 	{
 		GetArg(szArg1, szLine, 2);
-		gbShowRoleplaying = GetBoolFromString(szArg1, !gbShowRoleplaying);
+		gbShowAnon = GetBoolFromString(szArg1, !gbShowAnon);
 		WriteIni = true;
 	}
 	else if (ci_equals(szArg1, "sight"))
@@ -828,13 +828,13 @@ PLUGIN_API void OnPulse()
 						switch (pTarget->Anon)
 						{
 							case 1:
-								if (gbShowRoleplaying)
+								if (gbShowAnon)
 								{
 									strcpy_s(szTargetDist, "Anonymous");
 									break;
 								}
 							case 2:
-								if (gbShowRoleplaying)
+								if (gbShowAnon)
 								{
 									strcpy_s(szTargetDist, "Roleplaying");
 									break;